+++ libs/terraform/component/aws/src/lib/helpers/config.helper.ts
@@ -1,0 +1,8 @@
+import * as deepmerge from 'deepmerge';
+import { isPlainObject } from 'is-plain-object';
+
+export function mergeConfigurations<C, D>(defaultConfig: D, config: C): C & D {
+  return deepmerge(defaultConfig, config, {
+    isMergeableObject: isPlainObject,
+  }) as C & D;
+}