+++ libs/terraform/component/entrypoint/src/lib/index.ts
@@ -1,0 +1,2 @@
+export * from './entrypoint.component';
+export * from './entrypoint.interface';