+++ libs/terraform/component/s3/src/index.ts
@@ -1,1 +1,1 @@
-export * from './lib/terraform-s3-component';
+export * from './lib';