+++ libs/terraform/component/secrets/src/lib/secrets.interface.ts
@@ -1,0 +1,6 @@
+import { secretsmanager } from '@cdktf/provider-aws';
+
+export type SecretsComponentConfig = Record<string, never>;
+export interface SecretsComponentArtifacts {
+  secret: secretsmanager.SecretsmanagerSecret;
+}