+++ libs/CHANGELOG.md
@@ -2,6 +2,202 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.43.10](https://github.com/tractr/stack/compare/v1.43.9...v1.43.10) (2022-04-28)
+
+
+### Bug Fixes
+
+* **terraform:** change s3 versioning from Disabled to Suspended ([#421](https://github.com/tractr/stack/issues/421)) ([ddb85fe](https://github.com/tractr/stack/commit/ddb85fe3ffb6c9bab3df2d16fe185bd8d886f66b))
+
+
+
+## [1.43.10](https://github.com/tractr/stack/compare/v1.43.9...v1.43.10) (2022-04-28)
+
+
+### Bug Fixes
+
+* **terraform:** change s3 versioning from Disabled to Suspended ([#421](https://github.com/tractr/stack/issues/421)) ([ddb85fe](https://github.com/tractr/stack/commit/ddb85fe3ffb6c9bab3df2d16fe185bd8d886f66b))
+
+
+
+## [1.43.9](https://github.com/tractr/stack/compare/v1.43.8...v1.43.9) (2022-04-14)
+
+
+### Bug Fixes
+
+* add role check to react-admin ([#415](https://github.com/tractr/stack/issues/415)) ([b349e5a](https://github.com/tractr/stack/commit/b349e5a84ae6356a4c76fdbfd240709617e7fa39))
+
+
+
+## [1.43.9](https://github.com/tractr/stack/compare/v1.43.8...v1.43.9) (2022-04-14)
+
+
+### Bug Fixes
+
+* add role check to react-admin ([#415](https://github.com/tractr/stack/issues/415)) ([b349e5a](https://github.com/tractr/stack/commit/b349e5a84ae6356a4c76fdbfd240709617e7fa39))
+
+
+
+## [1.43.8](https://github.com/tractr/stack/compare/v1.43.7...v1.43.8) (2022-04-12)
+
+
+### Bug Fixes
+
+* edit result variable name in test ([#414](https://github.com/tractr/stack/issues/414)) ([c6d7e2e](https://github.com/tractr/stack/commit/c6d7e2ec7f87ce88294188bda4c9ed0f14bcd463))
+
+
+
+## [1.43.8](https://github.com/tractr/stack/compare/v1.43.7...v1.43.8) (2022-04-12)
+
+
+### Bug Fixes
+
+* edit result variable name in test ([#414](https://github.com/tractr/stack/issues/414)) ([c6d7e2e](https://github.com/tractr/stack/commit/c6d7e2ec7f87ce88294188bda4c9ed0f14bcd463))
+
+
+
+## [1.43.7](https://github.com/tractr/stack/compare/v1.43.6...v1.43.7) (2022-04-11)
+
+
+### Bug Fixes
+
+* **terraform-component-s3:** use name for bucket ([#413](https://github.com/tractr/stack/issues/413)) ([545a64e](https://github.com/tractr/stack/commit/545a64ec716e0bfde1452e62e560bac8e8fe4099))
+
+
+
+## [1.43.7](https://github.com/tractr/stack/compare/v1.43.6...v1.43.7) (2022-04-11)
+
+
+### Bug Fixes
+
+* **terraform-component-s3:** use name for bucket ([#413](https://github.com/tractr/stack/issues/413)) ([545a64e](https://github.com/tractr/stack/commit/545a64ec716e0bfde1452e62e560bac8e8fe4099))
+
+
+
+## [1.43.6](https://github.com/tractr/stack/compare/v1.43.5...v1.43.6) (2022-04-08)
+
+
+### Bug Fixes
+
+* **terraform-component-s3:** fix cors without rules ([#410](https://github.com/tractr/stack/issues/410)) ([a229d1e](https://github.com/tractr/stack/commit/a229d1e4533c5376c6515b012436ae58938025e6))
+
+
+
+## [1.43.6](https://github.com/tractr/stack/compare/v1.43.5...v1.43.6) (2022-04-08)
+
+
+### Bug Fixes
+
+* **terraform-component-s3:** fix cors without rules ([#410](https://github.com/tractr/stack/issues/410)) ([a229d1e](https://github.com/tractr/stack/commit/a229d1e4533c5376c6515b012436ae58938025e6))
+
+
+
+## [1.43.5](https://github.com/tractr/stack/compare/v1.43.4...v1.43.5) (2022-04-07)
+
+
+### Bug Fixes
+
+* update global guards to ignore rabbitmq requests ([#409](https://github.com/tractr/stack/issues/409)) ([e5d2b16](https://github.com/tractr/stack/commit/e5d2b169366abdc74d4a2cabcd9739201ccf76a1))
+
+
+
+## [1.43.5](https://github.com/tractr/stack/compare/v1.43.4...v1.43.5) (2022-04-07)
+
+
+### Bug Fixes
+
+* update global guards to ignore rabbitmq requests ([#409](https://github.com/tractr/stack/issues/409)) ([e5d2b16](https://github.com/tractr/stack/commit/e5d2b169366abdc74d4a2cabcd9739201ccf76a1))
+
+
+
+## [1.43.4](https://github.com/tractr/stack/compare/v1.43.3...v1.43.4) (2022-04-06)
+
+
+
+## [1.43.4](https://github.com/tractr/stack/compare/v1.43.3...v1.43.4) (2022-04-06)
+
+
+
+## [1.43.3](https://github.com/tractr/stack/compare/v1.43.2...v1.43.3) (2022-04-06)
+
+
+
+## [1.43.3](https://github.com/tractr/stack/compare/v1.43.2...v1.43.3) (2022-04-06)
+
+
+
+## [1.43.2](https://github.com/tractr/stack/compare/v1.43.1...v1.43.2) (2022-04-06)
+
+
+### Bug Fixes
+
+* **libs-terraform-group-network:** use interpolation for cidrsubnet ([#406](https://github.com/tractr/stack/issues/406)) ([6869498](https://github.com/tractr/stack/commit/68694984d02f2b4b841adeadc6bd70f11402a3f9))
+
+
+
+## [1.43.2](https://github.com/tractr/stack/compare/v1.43.1...v1.43.2) (2022-04-06)
+
+
+### Bug Fixes
+
+* **libs-terraform-group-network:** use interpolation for cidrsubnet ([#406](https://github.com/tractr/stack/issues/406)) ([6869498](https://github.com/tractr/stack/commit/68694984d02f2b4b841adeadc6bd70f11402a3f9))
+
+
+
+## [1.43.1](https://github.com/tractr/stack/compare/v1.43.0...v1.43.1) (2022-04-06)
+
+
+### Bug Fixes
+
+* **libs-terraform-service-ecs:** use last active task definition to detect last revision ([#405](https://github.com/tractr/stack/issues/405)) ([23a3629](https://github.com/tractr/stack/commit/23a362961a3ea37e5df7bf30c787f1201ec1f758))
+
+
+
+## [1.43.1](https://github.com/tractr/stack/compare/v1.43.0...v1.43.1) (2022-04-06)
+
+
+### Bug Fixes
+
+* **libs-terraform-service-ecs:** use last active task definition to detect last revision ([#405](https://github.com/tractr/stack/issues/405)) ([23a3629](https://github.com/tractr/stack/commit/23a362961a3ea37e5df7bf30c787f1201ec1f758))
+
+
+
+# [1.43.0](https://github.com/tractr/stack/compare/v1.42.1...v1.43.0) (2022-04-05)
+
+
+### Features
+
+* **libs-common:** array helpers and operators ([#404](https://github.com/tractr/stack/issues/404)) ([746c3fb](https://github.com/tractr/stack/commit/746c3fbe40b95fa96f56f969bb5780f559a4e9e3))
+
+
+
+# [1.43.0](https://github.com/tractr/stack/compare/v1.42.1...v1.43.0) (2022-04-05)
+
+
+### Features
+
+* **libs-common:** array helpers and operators ([#404](https://github.com/tractr/stack/issues/404)) ([746c3fb](https://github.com/tractr/stack/commit/746c3fbe40b95fa96f56f969bb5780f559a4e9e3))
+
+
+
+## [1.42.1](https://github.com/tractr/stack/compare/v1.42.0...v1.42.1) (2022-03-15)
+
+
+### Bug Fixes
+
+* add models exportation from models.module in nestjs-models templates ([#403](https://github.com/tractr/stack/issues/403)) ([9a78430](https://github.com/tractr/stack/commit/9a78430ab3d83b08db29682b0c229e6ec4a3d2cf))
+
+
+
+## [1.42.1](https://github.com/tractr/stack/compare/v1.42.0...v1.42.1) (2022-03-15)
+
+
+### Bug Fixes
+
+* add models exportation from models.module in nestjs-models templates ([#403](https://github.com/tractr/stack/issues/403)) ([9a78430](https://github.com/tractr/stack/commit/9a78430ab3d83b08db29682b0c229e6ec4a3d2cf))
+
+
+
 # [1.42.0](https://github.com/tractr/stack/compare/v1.41.0...v1.42.0) (2022-03-03)